Align harpguru-core with tutorial react-native "app"
There are things to take from the app currently named harpguru in this
project, and there are things to ignore from it. I need to update
several versions of things from harpguru-core to be in line with it, and
it looks advisable to bring some things across too. But not everything.
That app is also the expo boilerplate, and the harpguru-core package
doesn't need that stuff.