Remove v1 only reanimated usage
My current understanding of reanimated2 is that it still uses the
`Animated` default return object, but everything like `Node` & `Value`
that I was using before either aren't used anymore or are such low level
that they aren't going to be of use.

The strategy here is to strip down to a point where the application can
render again, and then build the animations back up again.