Add the HarpGuru component to the app & fix transpile
I'm don't really understand whether I *should* have the react-native
and reanimated babel transpilation config in both the boilerplate
package which loads the package which actually uses those packages, or
just in the boilerplate. I've convinced myself that since these packages
are only peer dependencies in the harpguru-core package that this
implies that it's not really the owner of the process and it's just a
library written with these packages included in their native form.

Probably I'll discover that the testing and linting will now start to
fail without a babel config to make sense of the typescript though.