Perform documented upgrade path from v1 to v2 reanimated
The previous commit proved that the app could still be basically
rendered by removing reanimated wherever it was erroring.

After I discovered the documentation on migrating from v1 to v2 I have
now reintroduced the files just as they are in the main harpguru
project, but taken the step of renaming `interpolate` and `Easing` to
`interpolateNode` and `EasingNode`.

That seems to have done a lot to fix things but there are still problems
which I'll have to investigate.

One main point to focus on initially is the removal of the
`react-native-redash` package. It's clear that this only works well with
v1 reanimated. I needed to remove the use of `interpolateColor` from
redash in this commit because it looked like it was using the old
`interpolate` from v1 and coming up against the new version of
`interpolate` from v2.

Even though there is no obvious v1 related error coming from redash now
I do think we need to just focus on getting rid of it and then seeing
where we are.