First attempt at replacing redash
This has not been a great success. There are lots of errors. However, I
do think it's some kind of foundation to build from, so I'm going to
take the next step by trimming back everything in the app to just a
single component and build back up from there.