Improve use of reanimated functions
Previously I don't think the values were actually resulting as hoped,
and the use of the provided hooks was much less standard. This is more
by the book at least.

I have not committed any change to harp-guru-page, but I have
temporarily deleted everything but the menus from there to achieve an
incrementally improvable lifecycle.