Convert root gesture handler syntax to v2
In theory this wasn't the type of commit that I was planning to do on
this branch. I had wanted to just replace the reanimated code and leave
the gesture handling code as it is until it's updated in a subsequent
branch. However, there actually isn't the backwards compatability that I
was expecting here, or I've just not figured out how to use reanimated 2
with gesture handler 1 and I'm having more luck by bringing gesture
handler 2 in to the mix, so I'm just going to proceed down this path of
least resistance.

It's not evident from this commit of course, but I've been experimenting
with the syntax and I'll try and push forward with what I've learned in
an incremental fashion.