Improve page drop animation
Springing is not ideal since it makes the overlaying comparison much
harder for the eye to do. Returning to interpolation also means that we
don't have to risk the offscreen page producing artifacts which effects
the onscreen page (because we can move the whole thing so much further
away offscreen now).