Set touches to pass through notification flash
This shouldn't be necessary *really* becuase the flash is set to animate
off the screen when it's done, but there seems to be a bug in iOS 16.1
which isn't there in 14.7.1 whereby the second page's notification flash
seems to fail to animate off screen as it's default position. There
needs to be a rerender triggered for it to find it's way to it's proper
place. But that isn't happening in all cases and it's blocking the
second page (invisibly).

I don't have time to investigate that bug when this perfectly good
workaround exists.