Ignore unresolvable eslint issues
Not necessarily unresolvable but not worth the effort of
resolving given the time constraint.

The object destructuring one would make the code less
readable so is not actually advised in this situation.

The generics one is not entirely clear but there is
some discussion about it being a bug which is yet to
be resolved:
https://github.com/typescript-eslint/typescript-eslint/issues/4062